+++ src/components/victory-line/line-segment.js
@@ -18,24 +18,21 @@ export default class LineSegment extends React.Component {
     return `curve${capitalize(interpolation)}`;
   }
 
+  renderLine(path, style, events) {
+    const { role } = this.props;
+    return (
+      <path style={style} d={path} role={role} {...events} vectorEffect="non-scaling-stroke"/>
+    );
+  }
+
   render() {
-    const { data, events, interpolation, role, scale, style } = this.props;
+    const { data, events, interpolation, scale, style } = this.props;
     const xScale = scale.x;
     const yScale = scale.y;
     const lineFunction = d3Shape.line()
       .curve(d3Shape[this.toNewName(interpolation)])
       .x((d) => xScale(d.x))
       .y((d) => yScale(d.y));
-    const path = lineFunction(data);
-
-    return (
-      <path
-        d={path}
-        {...events}
-        role={role}
-        style={style}
-        vectorEffect="non-scaling-stroke"
-      />
-    );
+    return this.renderLine(lineFunction(data), style, events);
   }
 }