+++ src/components/victory-line/victory-line.js
@@ -330,7 +330,13 @@ export default class VictoryLine extends React.Component {
     * @example theme={VictoryTheme.grayscale}
     * http://www.github.com/FormidableLabs/victory-core/tree/master/src/victory-theme/grayscale.js
     */
-    theme: PropTypes.object
+    theme: PropTypes.object,
+    /**
+     * The groupComponent prop takes an entire component which will be used to
+     * create group elements for use within container elements. This prop defaults
+     * to a <g> tag on web, and a react-native-svg <G> tag on mobile
+     */
+    groupComponent: PropTypes.element
   };
 
   static defaultProps = {
@@ -343,7 +349,8 @@ export default class VictoryLine extends React.Component {
     y: "y",
     dataComponent: <LineSegment/>,
     labelComponent: <VictoryLabel/>,
-    containerComponent: <VictoryContainer/>
+    containerComponent: <VictoryContainer/>,
+    groupComponent: <g/>
   };
 
   static getDomain = Domain.getDomain.bind(Domain);
@@ -375,7 +382,7 @@ export default class VictoryLine extends React.Component {
   }
 
   renderData(props) {
-    const { dataComponent, labelComponent } = props;
+    const { dataComponent, labelComponent, groupComponent } = props;
     const dataSegments = LineHelpers.getDataSegments(Data.getData(props));
 
     return dataSegments.map((data, key) => {
@@ -406,11 +413,9 @@ export default class VictoryLine extends React.Component {
         const lineLabel = React.cloneElement(labelComponent, Object.assign({
           events: Events.getPartialEvents(labelEvents, "all", labelProps)
         }, labelProps));
-        return (
-          <g key={`line-group-${key}`}>
-            {lineComponent}
-            {lineLabel}
-          </g>
+
+        return React.cloneElement(
+          groupComponent, {key: `line-group-${key}`}, lineComponent, lineLabel
         );
       }
       return lineComponent;
@@ -435,6 +440,14 @@ export default class VictoryLine extends React.Component {
     );
   }
 
+  renderGroup(children, style) {
+    return React.cloneElement(
+      this.props.groupComponent,
+      { role: "presentation", style},
+      children
+    );
+  }
+
   render() {
     const modifiedProps = Helpers.modifyProps(this.props, fallbackProps);
     const { animate, style, standalone } = modifiedProps;
@@ -449,7 +462,7 @@ export default class VictoryLine extends React.Component {
       ];
       return (
         <VictoryTransition animate={animate} animationWhitelist={whitelist}>
-          <VictoryLine {...modifiedProps}/>
+          {React.createElement(this.constructor, modifiedProps)}
         </VictoryTransition>
       );
     }
@@ -459,11 +472,7 @@ export default class VictoryLine extends React.Component {
 
     const baseStyles = Helpers.getStyles(style, styleObject, "auto", "100%");
 
-    const group = (
-      <g role="presentation" style={baseStyles.parent}>
-        {this.renderData(modifiedProps)}
-      </g>
-    );
+    const group = this.renderGroup(this.renderData(modifiedProps), baseStyles.parent);
 
     return standalone ? this.renderContainer(modifiedProps, group) : group;
   }